# src/zenml/zen_server/openapi3_0.yaml
openapi: 3.0.2
info:
  title: ZenML
  version: 0.xx.0
  description: >-
    The Zen API allows access to ZenML programmatically. These docs DO NOT represent the status quo but instead are the Design that is being aimed at.
...
paths:
...
  /projects:
    get:
      description: >-
        This call lists all projects in the organization based on the
        queried parameters.
      summary: List projects
      tags:
        - project
      operationId: get_projects_list
      responses:
        '200':
          description: Return all matched projects.
          content:
            application/json:
              schema:
                title: List of Projects
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    post:
      summary: Create Project
      tags:
        - project
      description: |-
        Creates a project based on the requestBody.
      operationId: create_project
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        '200':
          description: Returns the created Project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401': 
          $ref: '#/components/responses/401'
        '409': 
          $ref: '#/components/responses/409'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
  /projects/{project_name}:
    get:
      summary: Get a specific Project
      tags:
        - project
      description: |-
        Get a specific project using its unique name.
      operationId: get_project_by_name
      parameters:
        - $ref: '#/components/parameters/projectName'
      responses:
        '200':
          description: Returns the requested Project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'

      security:
        - HTTPBasic: []
    put:
      summary: Update a specific Project
      tags:
        - project
      description: |-
        Updates the attribute on a project.
      operationId: update_project_by_name
      parameters:
        - $ref: '#/components/parameters/projectName'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        '200':
          description: Returns the updated project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    delete:
      summary: Delete a specific Project
      tags:
        - project
      description: |-
        Deletes a project.
      operationId: delete_project_by_name
      parameters:
        - $ref: '#/components/parameters/projectName'
      responses:
        '200': 
          $ref: '#/components/responses/200'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
...
  /projects/{project_name}/pipelines:
    get:
      summary: Get pipelines that are defined within a project
      tags:
        - project
      description: |-
        Returns all pipelines within the project.
      operationId: get_pipelines_in_project
      parameters:
        - $ref: '#/components/parameters/projectName'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Stack Components in Project
                type: array
                items:
                  $ref: '#/components/schemas/Pipeline'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    post:
      summary: Create a pipeline within a project
      tags:
        - project
      description: |-
        Creates a pipeline based on the requestBody.
      operationId: create_pipeline
      parameters:
        - $ref: '#/components/parameters/projectName'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pipeline'
        required: true
      responses:
        '200':
          description: Returns the created Pipeline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
        '401': 
          $ref: '#/components/responses/401'
        '409': 
          $ref: '#/components/responses/409'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
...
components:
...
  schemas:
...
    Pipeline:
      title: Pipeline
      required:
        - name
        - id
      type: object
      properties:
        id:
          title: id
          type: string
          readOnly: True
          example: 9eaa2d6
        name:
          title: Name
          type: string
          example: cat_detection_training_pipeline
        description:
          title: Description
          type: string
          example: >-
            Pipeline to create a cat detector.
        status:
          $ref: '#/components/schemas/PreviousRunsStatus'
        owner:
          title: Owner
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string   
          readOnly: True
          description: >-
            User that created the trigger.
          example:
            {name: htahir, id: w813dw}
        created_at:
          title: Created at
          type: string
          format: date-time
          readOnly: True
          description: >-
            Time when this pipeline was first created.
          example: 2022-08-12T07:12:45.931Z
      description: |-
        A Pipeline is the blueprint for a collection of functions (or steps) 
        that are meant to fulfill one overarching task. 
        The flow of information through these functions is defined by a
        dag (directed acyclic graph).
...
    PreviousRunsStatus:
      title: Status' of the Previous Run
      type: object
      properties:
        1:
          $ref: '#/components/schemas/RunStatus'
        2:
          $ref: '#/components/schemas/RunStatus'
        3:
          $ref: '#/components/schemas/RunStatus'
    RunStatus:
      title: Run status
      type: object
      properties:
        status: 
          type: string
          enum:
            - Running
            - Failed
            - Finished
          readOnly: True
          example: Finished
        run:
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string
          description: >-
            The run.
...
    Project:
      title: Project
      required:
        - name
      type: object
      properties:
        id:
          title: Id
          type: string
          format: uuid
          readOnly: True
        created_at:
          title: Creation Date
          type: string
          format: date-time
          readOnly: True
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
      description: |-
        A Project is the encapsulation of all business use-case with all the required pipelines, infrastructure configurations and users with project-scoped permissions.
...
    ErrorModel:
      title: ErrorModel
      type: object
      properties:
        detail:
          title: Detail
      description: Base class for error responses.
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
  #-------------------------------
  # Reusable operation parameters
  #-------------------------------
  parameters:
...
    projectName:
      name: project_name
      in: path
      required: true
      schema:
        title: Project Name
        type: string
...
  responses:
    '200':
      description: Successful Response
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'        
    '409':
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    '422':
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
...
  securitySchemes:
    HTTPBasic:
      type: http
      scheme: basic